+++ components/script/dom/shadowroot.rs
@@ -16,7 +16,7 @@ use style::author_styles::AuthorStyles;
 use style::dom::TElement;
 use style::invalidation::element::restyle_hints::RestyleHint;
 use style::shared_lock::SharedRwLockReadGuard;
-use style::stylesheets::Stylesheet;
+use style::stylesheets::{DocumentStyleSheet, Stylesheet};
 use style::stylist::{CascadeData, Stylist};
 use stylo_atoms::Atom;
 
@@ -223,13 +223,19 @@ impl ShadowRoot {
             })
             .cloned();
 
-        self.document_or_shadow_root.add_stylesheet(
+        if self.document.has_browsing_context() {
+            let document_context = self.window.web_font_context();
+            self.window
+                .layout_mut()
+                .load_web_fonts_from_stylesheet(&sheet, &document_context);
+        }
+
+        DocumentOrShadowRoot::add_stylesheet(
             StylesheetSource::Element(Dom::from_ref(owner_node)),
             StylesheetSetRef::Author(stylesheets),
             sheet,
             insertion_point,
             self.document.style_shared_lock(),
-            self.document.has_browsing_context(),
         );
     }
 
@@ -243,24 +249,29 @@ impl ShadowRoot {
 
         let insertion_point = stylesheets.iter().last().cloned();
 
-        self.document_or_shadow_root.add_stylesheet(
+        if self.document.has_browsing_context() {
+            let document_context = self.window.web_font_context();
+            self.window
+                .layout_mut()
+                .load_web_fonts_from_stylesheet(&sheet, &document_context);
+        }
+
+        DocumentOrShadowRoot::add_stylesheet(
             StylesheetSource::Constructed(Dom::from_ref(cssom_stylesheet)),
             StylesheetSetRef::Author(stylesheets),
             sheet,
             insertion_point,
             self.document.style_shared_lock(),
-            self.document.has_browsing_context(),
         );
     }
 
     /// Remove a stylesheet owned by `owner` from the list of shadow root sheets.
     #[cfg_attr(crown, expect(crown::unrooted_must_root))] // Owner needs to be rooted already necessarily.
     pub(crate) fn remove_stylesheet(&self, owner: StylesheetSource, s: &Arc<Stylesheet>) {
-        self.document_or_shadow_root.remove_stylesheet(
+        DocumentOrShadowRoot::remove_stylesheet(
             owner,
             s,
             StylesheetSetRef::Author(&mut self.author_styles.borrow_mut().stylesheets),
-            self.document.has_browsing_context(),
         )
     }
